# Refactor detection FMetric