### Summary

Refactor FMetric by removing unused dynamic NMS threshold and optimizing the IOU/metric computations.

#### Evaluation Time Comparison
| Eval Period | Before PR Eval Time | PR Eval Time |
| :----: | :-------------------------:  |:------------------ |
|1 |146.7377 sec |  4.8786 sec |
|2 | 111.3974 sec | 4.1782 sec |
|3 | 52.1444 sec | 1.8198 sec |
|4 |103.8570 sec | 2.0841  sec |
|5 |68.9539 sec | 2.0969 sec | 
| 6|74.6123 sec | 2.0660 sec |

#### F1 Score Comparison
| Before PR test/F1 | PR test/F1 |
| :-------------------------:  |:------------------ |
|0.84606 |0.873132  |

#### Overall Elapsed Time
Before PR Elapsed time: 0:38:00.087019
PR Elapsed time: 0:13:25.230139

#### Note: 
* Two separate experiments were conducted using the YOLOX-s model on the same dataset, comparing results before and after refactor of the evaluation function.
* Although there is a difference in the F1 score (0.84606 vs. 0.873132) when tested, this discrepancy does not necessarily indicate an actual drop/increase in accuracy.
* When evaluating the same model before and after the refactor, the metric was identical (0.846069216), verifying that the refactor did not affect the model's performance.

### How to test

<!-- Describe the testing procedure for reviewers, if changes are
not fully covered by unit tests or manual testing can be complicated. -->

### Checklist

<!-- Put an 'x' in all the boxes that apply -->

- [x] I have added unit tests to cover my changes.​
- [x] I have added integration tests to cover my changes.​
- [ ] I have ran e2e tests and there is no issues.
- [x] I have added the description of my changes into CHANGELOG in my target branch (e.g., [CHANGELOG](https://github.com/openvinotoolkit/training_extensions/blob/develop/CHANGELOG.md) in develop).​
- [ ] I have updated the documentation in my target branch accordingly (e.g., [documentation](https://github.com/openvinotoolkit/training_extensions/tree/develop/docs) in develop).
- [x] I have [linked related issues](https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword).

### License

- [ ] I submit _my code changes_ under the same [Apache License](https://github.com/openvinotoolkit/training_extensions/blob/develop/LICENSE) that covers the project.
      Feel free to contact the maintainers if that's a concern.
- [ ] I have updated the license header for each file (see an example below).

```python
# Copyright (C) 2024 Intel Corporation
# SPDX-License-Identifier: Apache-2.0
```