@sovrasov 2024-11-25
@eugene123tw @kprokofi could you discuss the changes in details? From my side I'm afraid that this would affect the final estimated threshold value, which is important in case of small train/val. Perhaps, we should keep both slow/fast modes, but that depends on experiments. If the final confidence threshold estimation is now skewed, then it's ok to leave the fast version only.