@eugene123tw 2024-11-27
> @eugene123tw @kprokofi could you discuss the changes in details? From my side I'm afraid that this would affect the final estimated threshold value, which is important in case of small train/val. Perhaps, we should keep both slow (dynamic NMS threshold)/fast modes, but that depends on experiments. If the final confidence threshold estimation is now skewed, then it's ok to leave the fast version only.

@sovrasov @kprokofi refactoring the F1-score calculation is relatively safe. The primary change involves replacing the IoU computation loop with matrix operations, so making the evaluation more efficient and faster. The final F1-score results, including confidence thresholds, remain consistent.

I validated the before/after comparison across 9 datasets and there's no significant differences in accuracy. The results are summarized in the table below:

|      Model      | develop avg F1-score | PR avg F1-score |
|:---------------:|------------------|:---------------------------------:|
| ATSS_MOBILENET  | 0.7120           | 0.7067                            |
| ATSS_RESNEXT101 | 0.7284           | 0.7243                            |
| RTDETR_18       | 0.7221           | 0.7245                            |
| RTDETR_50       | 0.7418           | 0.7429                            |
| RTDETR_101      | 0.7378           | 0.7136                            |
| RTMDET_TINY     | 0.7107           | 0.7132                            |
| SSD_MOBILENET   | 0.5885           | 0.5818                            |
| YOLOX_L         | 0.7261           | 0.6633                            |
| YOLOX_S         | 0.7166           | 0.6923                            |
| YOLOX_TINY      | 0.6455           | 0.6454                            |
| YOLOX_X         | 0.7384           | 0.7381                            |
| YOLOV9_C        | 0.6696           | 0.6609                            |
| YOLOV9_M        | 0.6682           | 0.6663                            |
| YOLOV9_S        | 0.6726           | 0.6820                            |

Additionally, I verified functionality in tests/unit/core/metrics/test_fmeasure.py, and all checks passed without issues.