@eugene123tw 2024-11-28
> @eugene123tw could you have a look at iseg tests failure?
> 
> ```
>     results_per_confidence = self._get_results_per_confidence(
>   File "/home/validation/actions-runner/_work/training_extensions/training_extensions/.tox/integration-test-instance_segmentation/lib/python3.10/site-packages/otx/core/metrics/fmeasure.py", line 233, in _get_results_per_confidence
>     result_point = self.evaluate_classes(
>   File "/home/validation/actions-runner/_work/training_extensions/training_extensions/.tox/integration-test-instance_segmentation/lib/python3.10/site-packages/otx/core/metrics/fmeasure.py", line 279, in evaluate_classes
>     metrics, counters = self.get_f_measure_for_class(
>   File "/home/validation/actions-runner/_work/training_extensions/training_extensions/.tox/integration-test-instance_segmentation/lib/python3.10/site-packages/otx/core/metrics/fmeasure.py", line 323, in get_f_measure_for_class
>     batch_pred_bboxes = self.__filter_pred(
>   File "/home/validation/actions-runner/_work/training_extensions/training_extensions/.tox/integration-test-instance_segmentation/lib/python3.10/site-packages/otx/core/metrics/fmeasure.py", line 381, in __filter_pred
>     keep = (entity.labels == label_idx) & (entity.score > confidence_threshold)
> RuntimeError: The size of tensor a (66) must match the size of tensor b (100) at non-singleton dimension 0
> Traceback (most recent call last):
> ```

@sovrasov I forgot to filter scores in TV MaskRCNN post-processing. New change in this file: `src/otx/algo/instance_segmentation/segmentors/maskrcnn_tv.py`